% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggenrich.R
\name{keggenrich}
\alias{keggenrich}
\title{Perform KEGG Metabolite Enrichment Analysis}
\usage{
keggenrich(KEGGid, species, pathway_data, p.adjust.method = "BH")
}
\arguments{
\item{KEGGid}{Vector of metabolite KEGG IDs (e.g., "C00160")}

\item{species}{KEGG species code (e.g., "hsa")}

\item{pathway_data}{Pathway data obtained from getkpd()}

\item{p.adjust.method}{p-value adjustment method (default="BH")}
}
\value{
Enrichment results data frame
}
\description{
Conducts KEGG metabolite enrichment analysis using hypergeometric test
}
